refactor(cli): inline run-peer-errors into run.ts
The original split (`util/run-peer-errors.ts` ↔ test colocated in
`util/`) was justified in the spec as keeping error formatting
"unit-testable without standing up the full invoke pipeline." That
goal predates the `runCore`/`RunResult` refactor: today, `runCore`
returns typed `RunError` variants and `renderRunResult` is the only
caller of `emitMissError` / `emitRpcError`, so the formatters are
already isolated from the pipeline by the Result boundary — the
file split adds no extra testability.

Changes:
  - `emitMissError` and `emitRpcError` move into `run.ts`, exported.
  - The accompanying test file moves alongside `run.ts` and updates
    its import. Coverage is preserved — same 10 cases pass.
  - `util/run-peer-errors.ts` is deleted.

Net: one fewer file, one fewer cross-directory import, formatters
live next to their only call site.
